[deploy] Remove github_repos from public API (#6304)
