+++ app/controllers/github_repos_controller.rb
@@ -1,10 +1,26 @@
 class GithubReposController < ApplicationController
+  before_action :authenticate_user!
   after_action :verify_authorized
+
+  def index
+    authorize GithubRepo
+
+    client = create_octokit_client
+
+    existing_user_repos = current_user.github_repos.where(featured: true).
+      distinct.pluck(:github_id_code)
+
+    @repos = client.repositories.map do |repo|
+      repo.selected = existing_user_repos.include?(repo.id)
+      repo
+    end
+  rescue Octokit::Unauthorized => e
+    render json: { error: "Github Unauthorized: #{e.message}", status: 401 }, status: :unauthorized
+  end
 
   def create
     authorize GithubRepo
-    @client = create_octokit_client
-    @repo = GithubRepo.find_or_create(fetched_repo_params)
+    @repo = GithubRepo.find_or_create(fetched_repo_params(fetch_repo))
     current_user.touch(:github_repos_updated_at)
     if @repo.valid?
       redirect_to "/settings/integrations", notice: "GitHub repo added"
@@ -26,6 +42,27 @@ class GithubReposController < ApplicationController
     end
   end
 
+  def update_or_create
+    authorize GithubRepo
+
+    params[:github_repo] = JSON.parse(params[:github_repo])
+    fetched_repo = fetch_repo
+    unless fetched_repo
+      render json: "error: Could not find Github repo", status: :not_found
+      return
+    end
+
+    repo = GithubRepo.find_or_create(fetched_repo_params(fetched_repo))
+
+    current_user.touch(:github_repos_updated_at)
+
+    if repo.valid?
+      render json: { featured: repo.featured }
+    else
+      render json: "error: #{repo.errors.full_messages}"
+    end
+  end
+
   private
 
   def create_octokit_client
@@ -35,10 +72,7 @@ class GithubReposController < ApplicationController
     client
   end
 
-  def fetched_repo_params
-    fetched_repo = @client.repositories.detect do |repo|
-      repo.id == permitted_attributes(GithubRepo)[:github_id_code].to_i
-    end
+  def fetched_repo_params(fetched_repo)
     {
       github_id_code: fetched_repo.id,
       user_id: current_user.id,
@@ -53,5 +87,13 @@ class GithubReposController < ApplicationController
       featured: true,
       info_hash: fetched_repo.to_hash
     }
+  end
+
+  def fetch_repo
+    client = create_octokit_client
+
+    client.repositories.detect do |repo|
+      repo.id == permitted_attributes(GithubRepo)[:github_id_code].to_i
+    end
   end
 end
