+++ app/views/github_repos/index.json.jbuilder
