+++ config/routes.rb
@@ -133,11 +133,6 @@ Rails.application.routes.draw do
         get :organizations
         get :podcasts
       end
-      resources :github_repos, only: [:index] do
-        collection do
-          post "/update_or_create", to: "github_repos#update_or_create"
-        end
-      end
       resources :webhooks, only: %i[index create show destroy]
 
       resources :classified_listings, path: :listings, only: %i[index show create update]
@@ -185,7 +180,11 @@ Rails.application.routes.draw do
   resources :downloads, only: [:index]
   resources :stripe_active_cards, only: %i[create update destroy]
   resources :live_articles, only: [:index]
-  resources :github_repos, only: %i[create update]
+  resources :github_repos, only: %i[index create update] do
+    collection do
+      post "/update_or_create", to: "github_repos#update_or_create"
+    end
+  end
   resources :buffered_articles, only: [:index]
   resources :events, only: %i[index show]
   resources :additional_content_boxes, only: [:index]
